# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "デートプランを生成・管理するAPI",
        "title": "DatePlan API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "ヘルスチェック",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/plans": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plans"
                ],
                "summary": "デートプランを生成する",
                "parameters": [
                    {
                        "description": "プラン生成リクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PlanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザー一覧取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.UserResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザー作成",
                "parameters": [
                    {
                        "description": "ユーザー情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザー取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザー更新",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ユーザー情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "ユーザー削除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreatePlanRequest": {
            "type": "object",
            "required": [
                "prefecture"
            ],
            "properties": {
                "budget": {
                    "description": "予算 (default=Noneを表現するためポインタを使用)",
                    "type": "integer"
                },
                "date": {
                    "description": "いついくか",
                    "type": "string"
                },
                "desired_places": {
                    "description": "行きたい場所",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "end_time": {
                    "description": "終わりの時間 (default=Noneを表現するためポインタを使用)",
                    "type": "integer"
                },
                "has_car": {
                    "description": "車あり",
                    "type": "boolean"
                },
                "locations": {
                    "description": "場所(リスト),地名",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "prefecture": {
                    "description": "都道府県 (選ばせる項目)",
                    "type": "string"
                },
                "relationship": {
                    "description": "関係性",
                    "type": "string"
                },
                "start_time": {
                    "description": "始まりの時間",
                    "type": "integer"
                },
                "theme": {
                    "description": "テーマ",
                    "type": "string"
                },
                "time_slot": {
                    "description": "時間帯 (朝、昼、夜)",
                    "type": "string"
                }
            }
        },
        "dto.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.MovementInfo": {
            "type": "object",
            "properties": {
                "duration": {
                    "description": "時間(分)",
                    "type": "integer"
                },
                "from": {
                    "type": "string"
                },
                "method": {
                    "description": "徒歩 / 電車 / 車",
                    "type": "string"
                },
                "order": {
                    "description": "いつ表示するか",
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "dto.OpeningHours": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "integer"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
        "dto.PlanResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "movements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MovementInfo"
                    }
                },
                "spots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SpotInfo"
                    }
                },
                "theme": {
                    "type": "string"
                },
                "weather": {
                    "$ref": "#/definitions/dto.WeatherInfo"
                }
            }
        },
        "dto.SpotInfo": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "カフェ / 公園 / 映画館",
                    "type": "string"
                },
                "congestion": {
                    "description": "混雑度",
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "indoor_outdoor": {
                    "description": "屋内 / 屋外",
                    "type": "string"
                },
                "lat": {
                    "description": "緯度",
                    "type": "number"
                },
                "lng": {
                    "description": "経度",
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "opening_hours": {
                    "$ref": "#/definitions/dto.OpeningHours"
                },
                "order": {
                    "description": "いつ表示するか",
                    "type": "integer"
                },
                "photos": {
                    "description": "list[url]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price_range": {
                    "description": "価格帯",
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "stay_time": {
                    "description": "滞在時間目安(分)",
                    "type": "integer"
                }
            }
        },
        "dto.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.WeatherInfo": {
            "type": "object",
            "properties": {
                "season": {
                    "description": "季節",
                    "type": "string"
                },
                "status": {
                    "description": "晴れ / 雨 / 曇り",
                    "type": "string"
                },
                "temperature": {
                    "description": "気温",
                    "type": "number"
                }
            }
        }
    }
}